+++ source/Application/Controller/Admin/AdminlinksMain.php
@@ -7,6 +7,7 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
+use OxidEsales\Eshop\Core\TableViewNameGenerator;
 use oxRegistry;
 use stdClass;
 
@@ -28,7 +29,11 @@ class AdminlinksMain extends \OxidEsales\Eshop\Application\Controller\Admin\Admi
         parent::render();
 
         $soxId = $this->_aViewData["oxid"] = $this->getEditObjectId();
-        $oLinks = oxNew(\OxidEsales\Eshop\Application\Model\Links::class, getViewName('oxlinks'));
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $oLinks = oxNew(
+            \OxidEsales\Eshop\Application\Model\Links::class,
+            $tableViewNameGenerator->getViewName('oxlinks')
+        );
 
         if (isset($soxId) && $soxId != "-1") {
             $oLinks->loadInLang($this->_iEditLang, $soxId);
@@ -100,7 +105,11 @@ class AdminlinksMain extends \OxidEsales\Eshop\Application\Controller\Admin\Admi
         }
 
         $iEditLanguage = \OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter("editlanguage");
-        $oLinks = oxNew(\OxidEsales\Eshop\Application\Model\Links::class, getViewName('oxlinks'));
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $oLinks = oxNew(
+            \OxidEsales\Eshop\Application\Model\Links::class,
+            $tableViewNameGenerator->getViewName('oxlinks')
+        );
 
         if ($soxId != "-1") {
             //$oLinks->load( $soxId );
@@ -140,8 +149,11 @@ class AdminlinksMain extends \OxidEsales\Eshop\Application\Controller\Admin\Admi
         if (!isset($aParams['oxlinks__oxactive'])) {
             $aParams['oxlinks__oxactive'] = 0;
         }
-
-        $oLinks = oxNew(\OxidEsales\Eshop\Application\Model\Links::class, getViewName('oxlinks'));
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $oLinks = oxNew(
+            \OxidEsales\Eshop\Application\Model\Links::class,
+            $tableViewNameGenerator->getViewName('oxlinks')
+        );
         $iEditLanguage = \OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter("editlanguage");
 
         if ($soxId != "-1") {