+++ source/Application/Controller/Admin/DiscountMain.php
@@ -8,6 +8,7 @@
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
 use OxidEsales\Eshop\Core\Exception\InputException;
+use OxidEsales\Eshop\Core\TableViewNameGenerator;
 use oxRegistry;
 use oxDb;
 use stdClass;
@@ -93,7 +94,8 @@ class DiscountMain extends \OxidEsales\Eshop\Application\Controller\Admin\AdminD
         $sTitle = false;
         $sOxId = $this->getEditObjectId();
         if (isset($sOxId) && $sOxId != "-1") {
-            $sViewName = getViewName("oxarticles", $this->_iEditLang);
+            $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+            $sViewName = $tableViewNameGenerator->getViewName("oxarticles", $this->_iEditLang);
             // Reading from slave is ok here (see ESDEV-3804 and ESDEV-3822).
             $database = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
             $sQ = "select concat( $sViewName.oxartnum, ' ', $sViewName.oxtitle ) from oxdiscount