+++ source/Application/Controller/Admin/DynamicExportBaseController.php
@@ -7,6 +7,7 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
+use OxidEsales\Eshop\Core\TableViewNameGenerator;
 use oxRegistry;
 use oxDb;
 use stdClass;
@@ -352,8 +353,9 @@ class DynamicExportBaseController extends \OxidEsales\Eshop\Application\Controll
         $sLang = \OxidEsales\Eshop\Core\Registry::getLang()->getBaseLanguage();
         $oDB = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
 
-        $sCatView = getViewName('oxcategories', $sLang);
-        $sO2CView = getViewName('oxobject2category', $sLang);
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $sCatView = $tableViewNameGenerator->getViewName('oxcategories', $sLang);
+        $sO2CView = $tableViewNameGenerator->getViewName('oxobject2category', $sLang);
 
         //selecting category
         $sQ = "select $sCatView.oxleft, $sCatView.oxright, $sCatView.oxrootid from $sO2CView as oxobject2category left join $sCatView on $sCatView.oxid = oxobject2category.oxcatnid ";
@@ -399,8 +401,9 @@ class DynamicExportBaseController extends \OxidEsales\Eshop\Application\Controll
         $sLang = \OxidEsales\Eshop\Core\Registry::getLang()->getBaseLanguage();
         $oDB = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
 
-        $sCatView = getViewName('oxcategories', $sLang);
-        $sO2CView = getViewName('oxobject2category', $sLang);
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $sCatView = $tableViewNameGenerator->getViewName('oxcategories', $sLang);
+        $sO2CView = $tableViewNameGenerator->getViewName('oxobject2category', $sLang);
 
         //selecting category
         $sQ = "select $sCatView.oxtitle from $sO2CView as oxobject2category left join $sCatView on $sCatView.oxid = oxobject2category.oxcatnid ";
@@ -665,8 +668,9 @@ class DynamicExportBaseController extends \OxidEsales\Eshop\Application\Controll
         $oArticle = oxNew(\OxidEsales\Eshop\Application\Model\Article::class);
         $oArticle->setLanguage($iExpLang);
 
-        $sO2CView = getViewName('oxobject2category', $iExpLang);
-        $sArticleTable = getViewName("oxarticles", $iExpLang);
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $sO2CView = $tableViewNameGenerator->getViewName('oxobject2category', $iExpLang);
+        $sArticleTable = $tableViewNameGenerator->getViewName("oxarticles", $iExpLang);
 
         $insertQuery = "insert into {$sHeapTable} select {$sArticleTable}.oxid from {$sArticleTable}, {$sO2CView} as oxobject2category where ";
         $insertQuery .= $oArticle->getSqlActiveSnippet();
@@ -709,7 +713,8 @@ class DynamicExportBaseController extends \OxidEsales\Eshop\Application\Controll
     {
         if (!(\OxidEsales\Eshop\Core\Registry::getConfig()->getRequestParameter("blExportMainVars"))) {
             $oDB = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
-            $sArticleTable = getViewName('oxarticles');
+            $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+            $sArticleTable = $tableViewNameGenerator->getViewName('oxarticles');
 
             // we need to remove again parent articles so that we only have the variants itself
             $sQ = "select $sHeapTable.oxid from $sHeapTable, $sArticleTable where
@@ -791,7 +796,8 @@ class DynamicExportBaseController extends \OxidEsales\Eshop\Application\Controll
         if ($this->_aCatLvlCache === null) {
             $this->_aCatLvlCache = [];
 
-            $sCatView = getViewName('oxcategories');
+            $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+            $sCatView = $tableViewNameGenerator->getViewName('oxcategories');
             $oDb = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
 
             // Load all root cat's == all trees