+++ source/Application/Model/Search.php
@@ -7,6 +7,7 @@
 
 namespace OxidEsales\EshopCommunity\Application\Model;
 
+use OxidEsales\Eshop\Core\TableViewNameGenerator;
 use oxRegistry;
 use oxDb;
 
@@ -93,7 +94,8 @@ class Search extends \OxidEsales\Eshop\Core\Base
         $sSelect = $this->_getSearchSelect($sSearchParamForQuery, $sInitialSearchCat, $sInitialSearchVendor, $sInitialSearchManufacturer, false);
         if ($sSelect) {
             $sPartial = substr($sSelect, strpos($sSelect, ' from '));
-            $sSelect = "select count( " . getViewName('oxarticles', $this->_iLanguage) . ".oxid ) $sPartial ";
+            $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+            $sSelect = "select count( " . $tableViewNameGenerator->getViewName('oxarticles', $this->_iLanguage) . ".oxid ) $sPartial ";
 
             $iCnt = \OxidEsales\Eshop\Core\DatabaseProvider::getDb()->getOne($sSelect);
         }
@@ -185,8 +187,9 @@ class Search extends \OxidEsales\Eshop\Core\Base
         }
 
         $oArticle = oxNew(\OxidEsales\Eshop\Application\Model\Article::class);
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
         $sArticleTable = $oArticle->getViewName();
-        $sO2CView = getViewName('oxobject2category');
+        $sO2CView = $tableViewNameGenerator->getViewName('oxobject2category');
 
         $sSelectFields = $oArticle->getSelectFields();
 
@@ -198,7 +201,7 @@ class Search extends \OxidEsales\Eshop\Core\Base
 
         // must be additional conditions in select if searching in category
         if ($sInitialSearchCat) {
-            $sCatView = getViewName('oxcategories', $this->_iLanguage);
+            $sCatView = $tableViewNameGenerator->getViewName('oxcategories', $this->_iLanguage);
             $sInitialSearchCatQuoted = $oDb->quote($sInitialSearchCat);
             $sSelectCat = "select oxid from {$sCatView} where oxid = $sInitialSearchCatQuoted and (oxpricefrom != '0' or oxpriceto != 0)";
             if ($oDb->getOne($sSelectCat)) {
@@ -246,7 +249,8 @@ class Search extends \OxidEsales\Eshop\Core\Base
         $oDb = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
         $myConfig = \OxidEsales\Eshop\Core\Registry::getConfig();
         $blSep = false;
-        $sArticleTable = getViewName('oxarticles', $this->_iLanguage);
+        $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+        $sArticleTable = $tableViewNameGenerator->getViewName('oxarticles', $this->_iLanguage);
 
         $aSearchCols = $myConfig->getConfigParam('aSearchCols');
         if (!(is_array($aSearchCols) && count($aSearchCols))) {
@@ -310,7 +314,8 @@ class Search extends \OxidEsales\Eshop\Core\Base
         $searchColumns = \OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('aSearchCols');
 
         if (is_array($searchColumns) && in_array('oxlongdesc', $searchColumns)) {
-            $viewName = getViewName('oxartextends', $this->_iLanguage);
+            $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+            $viewName = $tableViewNameGenerator->getViewName('oxartextends', $this->_iLanguage);
             $descriptionJoin = " LEFT JOIN {$viewName } ON {$table}.oxid={$viewName }.oxid ";
         }
         return $descriptionJoin;
@@ -328,7 +333,8 @@ class Search extends \OxidEsales\Eshop\Core\Base
     protected function getSearchField($table, $field)
     {
         if ($field == 'oxlongdesc') {
-            $searchField = getViewName('oxartextends', $this->_iLanguage) . ".{$field}";
+            $tableViewNameGenerator = oxNew(TableViewNameGenerator::class);
+            $searchField = $tableViewNameGenerator->getViewName('oxartextends', $this->_iLanguage) . ".{$field}";
         } else {
             $searchField = "{$table}.{$field}";
         }